+++ CHANGELOG.md
@@ -53,7 +53,7 @@ Code Completion:
 Inlay Parameter Hints:
 
 1.  Don't display a hint when the argument variable name matches the parameter name.
-1.  Don't display a hint when the path step element name matches the parameter name.
+1.  Don't display a hint when the path step element/attribute name matches the parameter name.
 
 ## 2020